23/05/24

+++ lib/services/local_storage.dart
@@ -1,6 +1,9 @@
 import 'package:boton_ceti/animations/page_animation.dart';
+import 'package:boton_ceti/global/global_vars.dart';
+import 'package:boton_ceti/models/dynamic_alert_dialog.dart';
 import 'package:boton_ceti/views/login.dart';
 import 'package:flutter/material.dart';
+import 'package:lottie/lottie.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 
 class LocalStorage {
@@ -11,15 +14,85 @@ class LocalStorage {
   }
 
   static Future<void> clearStorage(BuildContext context) async {
-    await prefs.clear();
-    Future.microtask(
-      () => Navigator.of(context).pushAndRemoveUntil(
-          crearRutaNamed(const LoginScreen(), 'loginScreen'), (route) => false),
-    );
+    await Future.microtask(() => showDialog(
+          context: context,
+          builder: (context) {
+            return DynamicAlertDialog(
+              actions: [
+                ElevatedButton(
+                  onPressed: () => Future.microtask(
+                    () async {
+                      await prefs.clear();
+                      Future.microtask(
+                        () => Navigator.of(context).pushAndRemoveUntil(
+                            crearRutaNamed(const LoginScreen(), 'loginScreen'),
+                            (route) => false),
+                      );
+                    },
+                  ),
+                  style: ElevatedButton.styleFrom(
+                      shape: RoundedRectangleBorder(
+                        borderRadius: BorderRadius.circular(20),
+                      ),
+                      backgroundColor: VariablesGlobales.coloresApp[1]),
+                  child: const Text(
+                    'Confirmar',
+                    style: TextStyle(
+                      fontFamily: 'Nutmeg',
+                      fontWeight: FontWeight.w300,
+                    ),
+                  ),
+                )
+              ],
+              child: LayoutBuilder(
+                builder: (context, constraints) => Center(
+                  child: SizedBox(
+                    height: constraints.maxHeight * 0.9,
+                    width: constraints.maxWidth * 0.9,
+                    child: Column(
+                      mainAxisAlignment: MainAxisAlignment.center,
+                      children: [
+                        Lottie.asset(
+                          'assets/lotties/question.json',
+                          animate: true,
+                          repeat: false,
+                          fit: BoxFit.contain,
+                          frameRate: FrameRate(60),
+                        ),
+                        const SizedBox(height: 20),
+                        const Row(
+                          mainAxisAlignment: MainAxisAlignment.center,
+                          children: [
+                            Expanded(
+                              child: FittedBox(
+                                fit: BoxFit.scaleDown,
+                                child: Text(
+                                  '¿Quieres cerrar la sesión?',
+                                  style: TextStyle(
+                                    fontFamily: 'Nutmeg',
+                                    fontWeight: FontWeight.w300,
+                                  ),
+                                ),
+                              ),
+                            ),
+                          ],
+                        )
+                      ],
+                    ),
+                  ),
+                ),
+              ),
+            );
+          },
+        ));
   }
 
   static Future<void> clearUserProfiles() async {
     userProfiles = '';
+  }
+
+  static void changeLocationStatus() {
+    location = true;
   }
 
   static int? get idUsuario {
@@ -130,5 +203,27 @@ class LocalStorage {
 
   static set celular(String? val) {
     prefs.setString('celular', val!);
+  }
+
+  static bool? get location {
+    if (prefs.getBool('location') != null) {
+      return prefs.getBool('location');
+    }
+    return null;
+  }
+
+  static set location(bool? val) {
+    prefs.setBool('location', val!);
+  }
+
+  static bool? get locationEnabled {
+    if (prefs.getBool('location_enabled') != null) {
+      return prefs.getBool('location_enabled');
+    }
+    return null;
+  }
+
+  static set locationEnabled(bool? val) {
+    prefs.setBool('location_enabled', val!);
   }
 }
